Enable `make dev` from Outside Container
Running `make outside-dev` will now take care of running a container,
and building a development version of the code into
node_modules/ transparently. The library can be compiled in this way
without having purescript installed locally, and we can use a
host-system `node` repl to `require('Data.Luminous')` for manual
testing or exploration.